Change matrix element type from int to double
Replaces all int-based matrix storage and arithmetic with double to
support decimal values and avoid integer overflow in operations like
matrix multiply and scalar multiplication.

+++ main.c
@@ -14,7 +14,7 @@ struct my_struct {
     int id;            /* we'll use this field as the key */
     int column;
     int row;
-    int *matrix_mem;
+    double *matrix_mem;
 };
 struct my_struct *saved_matrices = NULL;
 
@@ -41,40 +41,40 @@ void delete_all_matrices()
     }
 }
 
-void printMatrixGlobalDim(int *given_matrix){//internally uses latest matrix dimensions
-    for(int i = 0; i < int_rows; i++){ 
+void printMatrixGlobalDim(double *given_matrix){//internally uses latest matrix dimensions
+    for(int i = 0; i < int_rows; i++){
        for(int j = 0; j < int_cols; j++){
-          printf("%d\t", given_matrix[i * int_cols + j]);
+          printf("%g\t", given_matrix[i * int_cols + j]);
        }
-       printf("\n"); 
+       printf("\n");
    }
 }
 
-void printMatrixGivenDim(int *fetched_matrix, int rows, int cols){//iterates based on given arguments
-    for(int i = 0; i < rows; i++){ 
+void printMatrixGivenDim(double *fetched_matrix, int rows, int cols){//iterates based on given arguments
+    for(int i = 0; i < rows; i++){
        for(int j = 0; j < cols; j++){
-          printf("%d\t", fetched_matrix[i * cols + j]);
+          printf("%g\t", fetched_matrix[i * cols + j]);
        }
-       printf("\n"); 
+       printf("\n");
    }
 }
 
-void setPointValue(int value, int row, int col, int rowSize, int colSize, int** matrix){
+void setPointValue(double value, int row, int col, int rowSize, int colSize, double** matrix){
   (*matrix)[row * int_cols + col] = value; //must pass dimensions rowSize colSize, otherwise C cannot determine where to place value as the array would then be varriable length array (VLA). So pointer arithmetic would fail
-					   //TODO prevent setting value out of bounds of array. C doesnt care if you mess up.
+				   //TODO prevent setting value out of bounds of array. C doesnt care if you mess up.
 }
 
-void fillValues(int value, int rows, int cols, int** matrix){
-  for (int i = 0; i < int_cols*int_rows; i++){ 
+void fillValues(double value, int rows, int cols, double** matrix){
+  for (int i = 0; i < int_cols*int_rows; i++){
 	   (*matrix)[i] = value;
-   } 
+   }
 }
 
 bool validateInput(char** input){
   return true;//TODO (0 < input || input  < int_rows -1) && (0 < input || input < int_cols-1);
 }
 
-void resizeMatrix(int rowDim, int colDim, int preRowDim, int preColDim, int** matrix){ //pass a pointer to the pointer, otherwise you will only modify local copy and caller points to old memory block, (make sure functions before this forward the pointer as ** otherwise you risk only having a copy of the pointer) 
+void resizeMatrix(int rowDim, int colDim, int preRowDim, int preColDim, double** matrix){ //pass a pointer to the pointer, otherwise you will only modify local copy and caller points to old memory block, (make sure functions before this forward the pointer as ** otherwise you risk only having a copy of the pointer)
 //If shrinking, it simply marks the unused tail as free
 //If moving, it copies the old data into a new smaller block
 //You never free individual elements
@@ -83,45 +83,45 @@ void resizeMatrix(int rowDim, int colDim, int preRowDim, int preColDim, int** ma
    int oldSize = preRowDim * preColDim;
    int newSize = rowDim * colDim;
 
-   *matrix = realloc(*matrix, newSize * sizeof(int));
+   *matrix = realloc(*matrix, newSize * sizeof(double));
  // Initialize new region
-   for (int i = oldSize; i < newSize; i++){ 
+   for (int i = oldSize; i < newSize; i++){
 	   (*matrix)[i] = 0;
    }
-   int_rows = rowDim; 
+   int_rows = rowDim;
    int_cols = colDim;
-   
+
 }
 
-void newRandomMatrix(int rowDim, int colDim, int preRowDim, int preColDim, int** matrix){
-   *matrix = realloc(*matrix, rowDim * colDim * sizeof(int));
+void newRandomMatrix(int rowDim, int colDim, int preRowDim, int preColDim, double** matrix){
+   *matrix = realloc(*matrix, rowDim * colDim * sizeof(double));
 
    srand(time(NULL));   // Initialization, should only be called once.
-   for (int i = 0; i < rowDim * colDim;  i++){ 
+   for (int i = 0; i < rowDim * colDim;  i++){
 	   (*matrix)[i] = rand();
    }
-   int_rows = rowDim; 
+   int_rows = rowDim;
    int_cols = colDim;
 }
 
-void multiplyMatrixWithInteger(int value, int rows, int cols, int** matrix){
-   for (int i = 0; i < int_cols*int_rows; i++){ 
+void multiplyMatrixWithInteger(double value, int rows, int cols, double** matrix){
+   for (int i = 0; i < int_cols*int_rows; i++){
 	   (*matrix)[i] = value * (*matrix)[i];
-   } 
+   }
 }
 
-int product(int a, int b){
-   int product = 0;
+double product(double a, double b){
+   double product = 0;
    product = a * b;
    return product;
 }
 
-int* iterateSumProductOverSndMatrixCol(int* fstMatrRow, int n1, int** matrix, int k, int n2, int* sequenceOfDotProducts){
- int sndMatrixCol[n2];
+double* iterateSumProductOverSndMatrixCol(double* fstMatrRow, int n1, double** matrix, int k, int n2, double* sequenceOfDotProducts){
+ double sndMatrixCol[n2];
 
-int dot_product = 0;
+double dot_product = 0;
 
-sequenceOfDotProducts = realloc(sequenceOfDotProducts, (k) * sizeof(int) );
+sequenceOfDotProducts = realloc(sequenceOfDotProducts, (k) * sizeof(double) );
 
  for (int i = 0; i < k; i++){
     int row = 0;
@@ -142,19 +142,19 @@ return sequenceOfDotProducts;
 }
 
 //transpose current matrix, swapping rows and columns
-void transposeMatrix(int** matrix){
+void transposeMatrix(double** matrix){
    int new_rows = int_cols;
    int new_cols = int_rows;
 
-   int *result = malloc(new_rows * new_cols * sizeof(int));
+   double *result = malloc(new_rows * new_cols * sizeof(double));
    for (int i = 0; i < int_rows; i++){
       for (int j = 0; j < int_cols; j++){
          result[j * new_cols + i] = (*matrix)[i * int_cols + j];
       }
    }
 
-   *matrix = realloc(*matrix, new_rows * new_cols * sizeof(int));
-   memcpy(*matrix, result, new_rows * new_cols * sizeof(int));
+   *matrix = realloc(*matrix, new_rows * new_cols * sizeof(double));
+   memcpy(*matrix, result, new_rows * new_cols * sizeof(double));
    free(result);
 
    int_rows = new_rows;
@@ -162,26 +162,26 @@ void transposeMatrix(int** matrix){
 }
 
 //overwrite current matrix with identity matrix, only valid for square matrices
-void generateIdentityMatrix(int** matrix){
+void generateIdentityMatrix(double** matrix){
    if (int_rows != int_cols){
       printf("Cannot generate identity: matrix is not square (%dx%d)\n", int_rows, int_cols);
       return;
    }
    for (int i = 0; i < int_rows; i++){
       for (int j = 0; j < int_cols; j++){
-         (*matrix)[i * int_cols + j] = (i == j) ? 1 : 0;
+         (*matrix)[i * int_cols + j] = (i == j) ? 1.0 : 0.0;
       }
    }
 }
 
 //add saved matrix to current matrix, both must have equal dimensions
-void addMatrixToMatrix(int matrixId, int** matrix){
+void addMatrixToMatrix(int matrixId, double** matrix){
    struct my_struct *thisMatrix = find_matrix(matrixId);
    if (thisMatrix == NULL){
       printf("Matrix with id %d not found\n", matrixId);
       return;
    }
-   int *matrix_mem = thisMatrix -> matrix_mem;
+   double *matrix_mem = thisMatrix -> matrix_mem;
    int rows = thisMatrix -> row;
    int cols = thisMatrix -> column;
 
@@ -196,13 +196,13 @@ void addMatrixToMatrix(int matrixId, int** matrix){
 }
 
 //subtract saved matrix from current matrix, both must have equal dimensions
-void subtractMatrixFromMatrix(int matrixId, int** matrix){
+void subtractMatrixFromMatrix(int matrixId, double** matrix){
    struct my_struct *thisMatrix = find_matrix(matrixId);
    if (thisMatrix == NULL){
       printf("Matrix with id %d not found\n", matrixId);
       return;
    }
-   int *matrix_mem = thisMatrix -> matrix_mem;
+   double *matrix_mem = thisMatrix -> matrix_mem;
    int rows = thisMatrix -> row;
    int cols = thisMatrix -> column;
 
@@ -217,18 +217,18 @@ void subtractMatrixFromMatrix(int matrixId, int** matrix){
 }
 
 //multiply matrix a (m x n) with matrix b (n x k), so the product is m x k
-void multiplyMatrixWithMatrix(int matrixId, int grow, int gcols, int** matrix){
-   
+void multiplyMatrixWithMatrix(int matrixId, int grow, int gcols, double** matrix){
+
    struct my_struct *thisMatrixToSave = find_matrix(matrixId);
    if (thisMatrixToSave == NULL){
       printf("Matrix with id %d not found\n", matrixId);
       return;
    }
-   int *matrix_mem = thisMatrixToSave -> matrix_mem;
+   double *matrix_mem = thisMatrixToSave -> matrix_mem;
    int rows = thisMatrixToSave -> row;
    int cols = thisMatrixToSave -> column;
 
-   //first: m x n1, 
+   //first: m x n1,
    int m = rows;
    int n1 = cols;
    printf("m: %d, n1: %d\n", m, n1);
@@ -237,20 +237,20 @@ void multiplyMatrixWithMatrix(int matrixId, int grow, int gcols, int** matrix){
    int n2 = int_rows;
    int k = int_cols;
    printf("n2: %d, k: %d\n", n2, k);
-   
+
    if (n1 != n2){
       printf("column and row does not match between both matrices, can't calculate\n");
       return;
    }
 
    //new matrix m x k
-   int *result_matrix = malloc(m * k * sizeof(int)); //flat buffer/single contiguous block of memory
+   double *result_matrix = malloc(m * k * sizeof(double)); //flat buffer/single contiguous block of memory
 
- int fstMatrixRow[n1];
+ double fstMatrixRow[n1];
 
  int offsetRow = 0;
 
-int* result = malloc(k * sizeof(int));
+double* result = malloc(k * sizeof(double));
 int offsetResultIndex = 0;
  for (int i = 0; i < m * n1; i=i+n1){
     offsetRow = i;
@@ -264,19 +264,19 @@ int offsetResultIndex = 0;
       offsetResultIndex += k;
  }
 
-*matrix = realloc(*matrix, m * k *sizeof(int));
+*matrix = realloc(*matrix, m * k * sizeof(double));
 for (int i = 0; i < m * k; i++){
    (*matrix)[i] = result_matrix[i];
 }
 
-   int_rows = m; 
+   int_rows = m;
    int_cols = k;
    free(result);
    free(result_matrix);
 }
 
-void doActionInput(char** action, int rowSize, int colSize, int* matrix, int** matrixp){
-   
+void doActionInput(char** action, int rowSize, int colSize, double* matrix, double** matrixp){
+
    if (strcmp(action[0], "exit") == 0){
       free(matrix);
       free(action[0]);
@@ -285,69 +285,69 @@ void doActionInput(char** action, int rowSize, int colSize, int* matrix, int** m
       exit(EXIT_SUCCESS);
    } else if (strcmp(action[0], "set") == 0){
      //> set 2 2 7 //set value in row 2 column 2 to value 7
-       setPointValue(atoi(action[3]), atoi(action[1]), atoi(action[2]), int_rows, int_cols, &matrix);
-       free(action[0]); 
+       setPointValue(atof(action[3]), atoi(action[1]), atoi(action[2]), int_rows, int_cols, &matrix);
+       free(action[0]);
        free(action[1]);
        free(action[2]);
-       free(action[3]);  
+       free(action[3]);
    } else if (strcmp(action[0], "fill") == 0){
-     //> fill 3 //fill entire matrix with value 3 
-      fillValues(atoi(action[1]), int_rows, int_cols, &matrix);
-      free(action[0]); 
+     //> fill 3 //fill entire matrix with value 3
+      fillValues(atof(action[1]), int_rows, int_cols, &matrix);
+      free(action[0]);
       free(action[1]);
    } else if (strcmp(action[0], "resize") == 0){
      //> resize 4 3 //resize existing matrix to 4 x 3
       resizeMatrix(atoi(action[1]), atoi(action[2]), int_rows, int_cols, matrixp);
-      free(action[0]); 
+      free(action[0]);
       free(action[1]);
       free(action[2]);
    } else if (strcmp(action[0], "newrand") == 0){
      //> newrand 4 3 //make new 4 x 3 matrix with random values
      newRandomMatrix(atoi(action[1]), atoi(action[2]), int_rows, int_cols, matrixp);
-     free(action[0]); 
+     free(action[0]);
      free(action[1]);
      free(action[2]);
    } else if (strcmp(action[0], "mult") == 0){
      //> mult 2 //multiply each value in matrix by 2
-     multiplyMatrixWithInteger(atoi(action[1]), int_rows, int_cols, &matrix);
-     free(action[0]); 
+     multiplyMatrixWithInteger(atof(action[1]), int_rows, int_cols, &matrix);
+     free(action[0]);
      free(action[1]);
    } else if (strcmp(action[0], "save") == 0){
      //> save 2 //save matrix with with key 2
-     
+
      struct my_struct *thisMatrixToSave = malloc(sizeof(struct my_struct));
-     thisMatrixToSave -> matrix_mem = malloc(int_rows * int_cols * sizeof(int));
+     thisMatrixToSave -> matrix_mem = malloc(int_rows * int_cols * sizeof(double));
      thisMatrixToSave -> id = atoi(action[1]);
      thisMatrixToSave -> column = int_cols;
      thisMatrixToSave -> row = int_rows;
-    
-     memcpy(thisMatrixToSave -> matrix_mem, matrix, int_rows * int_cols * sizeof(int));
+
+     memcpy(thisMatrixToSave -> matrix_mem, matrix, int_rows * int_cols * sizeof(double));
 
      save_matrix(thisMatrixToSave);
-     free(action[0]); 
+     free(action[0]);
      free(action[1]);
 
    } else if (strcmp(action[0], "load") == 0){
      //> load 2 //find the saved matrix by looking up key 2
      struct my_struct *thisMatrixToSave = find_matrix(atoi(action[1]));
-     int *matrix_mem = thisMatrixToSave -> matrix_mem;
+     double *matrix_mem = thisMatrixToSave -> matrix_mem;
      int rows = thisMatrixToSave -> row;
      int cols = thisMatrixToSave -> column;
-   
+
      if (int_rows * int_cols < rows * cols ){
         //works when loading big matrix to small current matrix
-        *matrixp = realloc(*matrixp, rows * cols * sizeof(int));
-	for (int i = 0; i < rows * cols;  i++){ 
+        *matrixp = realloc(*matrixp, rows * cols * sizeof(double));
+	for (int i = 0; i < rows * cols;  i++){
 	   (*matrixp)[i] = matrix_mem[i];
         }
      } else {
-        memcpy(matrix, matrix_mem, rows * cols * sizeof(int)); 
+        memcpy(matrix, matrix_mem, rows * cols * sizeof(double));
      }
 
      int_rows = rows;
      int_cols = cols;
 
-     free(action[0]); 
+     free(action[0]);
      free(action[1]);
 
    } else if (strcmp(action[0], "mm") == 0){
@@ -379,23 +379,23 @@ void doActionInput(char** action, int rowSize, int colSize, int* matrix, int** m
 
 }
 
-void parseProgramInput(char** in, int rowSize, int colSize, int* matrix, int** matrixp){
+void parseProgramInput(char** in, int rowSize, int colSize, double* matrix, double** matrixp){
   //TODO validate arg 'in' along with dimensions rowSize and colSize
 
   if (validateInput(in) == false){
      printf("Error incorrect input or dimension \n");
      return;
   }
-  doActionInput(in, int_rows, int_cols, matrix, matrixp);   
+  doActionInput(in, int_rows, int_cols, matrix, matrixp);
 }
 
-char** tokenize(char in[]){ 
+char** tokenize(char in[]){
     char* temp = strdup(in);
     char* token = strtok(temp, " ");
 
     char** list = NULL;
     int count = 0;
-    
+
     while(token != NULL){
 
     list = realloc(list, (count + 1) * sizeof(char*));
@@ -412,9 +412,9 @@ int main(int argc, char *argv[]){
    int_rows = *argv[1] - '0'; //first matrix dimension program arg
    int_cols = *argv[2] - '0'; //second matrix dimension program arg
 
-   int *matrix = malloc(int_rows * int_cols * sizeof(int)); //flat buffer/single contiguous block of memory
+   double *matrix = malloc(int_rows * int_cols * sizeof(double)); //flat buffer/single contiguous block of memory
    for (int i = 0; i < int_rows * int_cols; i++){
-	   matrix[i] = 0;
+	   matrix[i] = 0.0;
   }
    printf("This is graph program: Hello %s \n","User");
    printf("The Graph dimension is %i x %i\n", int_rows, int_cols);
@@ -423,7 +423,7 @@ int main(int argc, char *argv[]){
    printf("> resize 4 3 //make new 4 x 3 matrix\n");
    printf("> newrand 4 3 //make new 4 x 3 matrix with random values\n");
    printf("> mult 2 //multiply each value in matrix by 2\n");
-   printf("> save 2 //save matrix with with key 2\n"); 
+   printf("> save 2 //save matrix with with key 2\n");
    printf("> load 2 //find the saved matrix by looking up key 2\n");
    printf("> mm 2 //find the saved matrix by looking up key 2, then multiply it with matrix in current context\n");
    printf("> tr //transpose current matrix\n");
@@ -431,7 +431,7 @@ int main(int argc, char *argv[]){
    printf("> ma 2 //add saved matrix with key 2 to the current matrix (dimensions must match)\n");
    printf("> ms 2 //subtract saved matrix with key 2 from the current matrix (dimensions must match)\n");
    while(1){
-    printMatrixGlobalDim(matrix);    
+    printMatrixGlobalDim(matrix);
 
     char in[100];
     scanf(" %[^\n]", in);
@@ -439,7 +439,7 @@ int main(int argc, char *argv[]){
     char** list = tokenize(in);
 
     parseProgramInput(list, int_rows, int_cols, matrix, &matrix);
-        
+
     free(list);
     }
 }
